requirements.txt
confluent-kafka==2.5.3
simplejson==3.19.3
Django==4.2.16
psycopg2-binary>=2.8
djangorestframework==3.12.4
channels==3.0.4
channels-redis==3.3.1